// src/app/model/table-model.ts
export default class TableModel {
    iterationcount: number;
    equipments: number;
    lines:number;
    numberOfObjects: number;
    numberOfRulesChecked: number;
    equivalency: string;
    isTimedOut: string;
    equipmentPlacementTime: number;
    pipeRouterTime: number;
    equivalencyTime: number;
    totalElpsedTime: number;
    throughput: number;
  }